requirements.txt
pydicom==2.4.0
pydicom-seg==0.2.3
pytorch-lightning==1.9.0
timm==1.0.16
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.3.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.43.3
ultralytics==8.3.161
nltk==3.9.1
numpy==1.26.4
einops==0.8.1
monai==1.5.0
open-clip-torch==2.32.0
opencv-contrib-python==4.11.0.86
ml_collections==1.1.0
scikit-learn==1.5.1
